refactored experience component, moved ShowAddExperience component into separate file
